move useSuspense into a hook
