+++ packages/extras/src/lib/components/BakeShadows/BakeShadows.svelte
@@ -1,25 +1,5 @@
 <script lang="ts">
-  import { useThrelte } from '@threlte/core'
-  import { useSuspense } from '@threlte/extras'
-  import { fromStore } from 'svelte/store'
+  import { useSuspense } from '$lib/suspense/useSuspense.js'
 
-  const suspended = fromStore(useSuspense().suspended)
-
-  const { renderer } = useThrelte()
-
-  $effect.pre(() => {
-    if (suspended.current) {
-      return
-    }
-
-    const { autoUpdate } = renderer.shadowMap
-
-    renderer.shadowMap.autoUpdate = false
-    renderer.shadowMap.needsUpdate = true
-
-    return () => {
-      renderer.shadowMap.autoUpdate = autoUpdate
-      renderer.shadowMap.needsUpdate = true
-    }
-  })
+  useSuspense()
 </script>
